[SDA-3063] ClusterDetails: Fix a crash when installing an OSD cluster
Looks like specifying the dispatch explicitly in `mapDispatchToProps`
for support contacts caused it to be undefined when it was called via
`<ClusterStatusMonitor>` -> `refresh()`.

Keeping dispatch implicit there seems to fix this issue. I also added a
condition to avoid fetching this data when the feature is not enabled.

Should fix https://issues.redhat.com/browse/SDA-3063