[SDA-3061] Don't use metrics for OSD upgrade availability
Some versions are disabled for OSD, so using metrics to check if an
update is available is problematic. For now, the update button will be
removed from the cluster list, and will only appear in the cluster
details screen. To check if an update is available, we query the version
information for the specific cluster version, and check if the backend
knows about any available upgrades.

Even if available upgrades becomes a property of the cluster object
itself this new component `<ClusterVersionInfo>` will be useful, since
it'll soon (in a followup MR probably) gain the ability to also query
the upgrade schedule, to indicate if an update is scheduled and allow
the user to cancel it.