Upgrades: Select time in local time instead of UTC
To reduce the chances of weird timezone-related edge cases and to be
less confusing for users, select time in UTC and show the current time
below.

Also scroll to the selected time when the time selection is
opened, and disable anything later than the default timestamp which
should be one hour from the current time, rounded to the closest 15
minutes increment.

And fix the spacing between the title and the grid in the version
selection grid.

Should fix:
 - SDA-3175
 - SDA-3192