Refactor modal usage
This is a refactor that takes handling `isOpen` away
from the individual modals into a wrapper that handles it, and define
modal names in consts rather than repeating a string everywhere.

This reduces the complexity of writing modal code that needs to refresh
resources from the backend when it's opened - instead of manually
checking `isOpen` in `componentDidUpdate` - you can just rely on
`componentDidMount`, as the modal will be unmounted when it's closed.
For modals with internal state it'll also mean that we could skip
resetting the state on close, because state resets when a component is
unmounted.

In addition, a new component `<CommonClusterModals />` was added, to
unify modals shared between Cluster Details and Cluster List, reducing
duplication and making it easier to correctly add a new modal.