SDA-3337 - Monitoring: use OSD upgrade policy state for "Upgrading"
This MR changes MonitoringSelectors to use the upgrade policy state to
check if an OSD cluster is upgrading. As long as the state value is
`started` we consider the upgrade as if it's in progress. These updates
take longer than what metrics report (as far as I can tell, though I
don't know why) and start before metrics about it are sent.