SDA-3302: upgrade settings - reset schedule when switching to manual
SDA-3302 describes a case where if you have a cluster that is set for
manual updates and you go into the upgrade settings tab, select
"automatic", change the schedule, and then change your mind and go back
to "manual", the "save" button remained enabled.

This change fixes this bug by changing the value of the schedule back to
the default value, making redux-form consider the form pristine in this
case.