SDA-3442: CreateOSD: unify heading sizes, remove .pf-c-content
The `pf-c-content` class on the form adds a margin to every `<ul>`,
making any <Select> render the options slightly mis-aligned to the
select toggle. This was reported in SDA-3442

The best way to fix that is to remove `pf-c-content`, but that made all
headings look like normal text, so I've taken the opportunity to make
sure all titles on the form use a pattenfly-react `<Title>` with a
consistent heading level / size.