NetworkSelfService: Fix test for "should open modal"
Use jest fake timers correctly.

Opening the modal with a timeout is done as a workaround to a PF bug,
see https://github.com/patternfly/patternfly-react/issues/5183 for more
details.

It was merged in
https://gitlab.cee.redhat.com/service/uhc-portal/-/merge_requests/1910
with an incorrect use of setTimeout in the test.

This MR is a followup that fixes the test.