SDA-3521 - Fix the EditSubscriptionSettingsDialog
This modal was broken after my modal refactors, because it relies on
converting props into state. It only had it in the
componentWillRecieveProps lifecycle method, but now that it's using the
ConnectedModal wrapper it gets these props on mount.

I didn't remove the old lifecycle method to avoid causing
difficult-to-debug breakage. This modal urgently needs to be refactored
to fit in with the rest of our codebase, instead of doing its' own
special thing