SDB-1730: Fix consoleURL discrepancy in cluster details
https://gitlab.cee.redhat.com/service/uhc-portal/-/merge_requests/2024
was the right approach, but it was missing `.data`, and accessed
`cluster` instead of `cluster.data`, so the info was put on the request
status object and not its actual data.

Also fixed a case where openshift_version existed in AMS metrics, but
not in the CS cluster resource.

This is all temporary until the clusterDetails refactor gets merged.