sentry: Record destroyed form in redux breadcrumbs
There are some issues in Sentry which seem to happen when a specific
form gets destroyed/removed, such as https://sentry.devshift.net/sentry/ocm-ui/issues/263019/?referrer=AssignedActivityEmail

while these errors are (probably) harmless, it'd be a good idea to
figure out why they happen. As this doesn't happen every time, and it
happens after a form gets "destroyed", we should log the forms getting
"destroyed" in the breadcrumbs, so we have more insights as to when
exactly this issue occurs.