SDA-3518: Cluster Details - use Subscription as the main resource
This MR is the first step in a refactor that'll make Subscription the
base resource, and `cluster` (from Clusters Service) an additional-info
record that will only be relevant for OSD/ROSA clusters.

The end goal is to remove OCP clusters from Clusters Service, and to
allow Assisted Installer clusters to integrate more cleanly into OCM.

In a followup MR the "fake cluster record" generated from the
subscription for compatibility reasons will be removed, so
`subscription` will be the main record, and OSD/ROSA specific data will
be in `subscription.managedCluster`.