[SDA-3500] Upgrades: use available_upgrades from cluster object
Using available_upgrades from the cluster object allows us to delete a
lot of code, and also properly support ROSA clusters.

The available_upgrades field is just an array of version numbers, and
querying each of them to check its' rosa_enabled status didn't really
scale. Now that the available_upgrades is on the cluster object, we get
a pre-filtered list of updates that are enabled for the specific type of
cluster.

While in theory this would've allowed us to also add the update link to
the cluster list, but in practice we can't check if an update is
scheduled when we're on the list, so the button is still hidden there
for OSD clusters.

Fixes:
https://issues.redhat.com/browse/SDA-3500
https://issues.redhat.com/browse/SDA-3180