switch from node-sass to sass
node-sass is an implementation of the sass (scss) preprocessor that uses
a natively-compiled C library (libSass) to convert scss to css.

sass is a pure JS implementation doing the same thing.

while node-sass is faster than sass, it has the disadvantage of
requiring compilation of native code, which doesn't always work and
usually breaks when people update nodejs or for new developers starting
out with OCM.

our current version of node-sass is incompatible with the most recent
version of node, but to update it we'll also need to update webpack,
which is a lot more work due to breaking changes in webpack 5.0.

Switching to sass prevents this problem, as it doesn't require
compilation of native code it's less likely to break and more likely to
work out of the box for everyone. A few extra seconds of build time is a
fair price to pay for this benefit in my opinion.