SDA-3751: UpgradeSettingsTab - hide update button when version is unknown
When the cluster is installing, before we got metrics,
cluster.openshift_version is not defined. In general, it doesn't make
sense for the user to be able to schedule an update before the cluster
is ready. This commit makes sure the button won't be shown when the
version is unknown.

fixes https://issues.redhat.com/browse/SDA-3751

Also fixes capitalization - https://issues.redhat.com/browse/SDA-3751