Support quota user for Selenium, remove backned proxy from tests
This commits add supports for the new with-quota user for our Selenium
tests. For now, for simplicity, I changed all wdio to use the with-quota
user.

This depends on https://gitlab.cee.redhat.com/service/app-interface/-/merge_requests/15594

Additionally, I've removed the use of the backend-proxy and qa.foo from
Selenium tests. Since selenium tests can already use the more stable
prod.foo, and prod.foo gives us a real, usable API token, we can skip
the backend proxy entirely, making the test process simpler and more
similar to a real user.

When running locally with a running full `yarn start`
development environment, `yarn run wdio` will continue using qa.foo,
this simplifies development of tests as you don't have to export the
env vars with the credentials.

To achieve bypassing of the `./backend` proxy when running locally, I've
changed the logic in `config.js` to respect the API_ENV flag from
webpack even in development mode, and moved default-selection logic into
weback - if the `--api-env` param is specified, it'll be used no matter
if you're using dev server or a real build. Otherwise, it'll default to
`development` config (backend proxy) when running in a dev server, and
to `production` config when it's a real build.