SDA-3825: Fake "deleting" status for an AWS role grant when deleting
When we delete a role grant, we wait for the request to be fulfilled,
then we fetch grants again. On fast networks this happens quickly
enough, but on slow network (like the one QE uses) it takes a while.
This commits keeps the ID of the deleted row in the state, so we can
fake a "deleting" status until we fetch the list again for an updated
status from the server.

Should fix SDA-3825, at least partially. The other part of this issue is
asking for some kind of indication fetch is in progress after
deletion (like the other sections),
but this is a conflicting ask with faking the Deleting status.
we can do either of these - not both!

Unlike other sections, role grants can take a while to delete (when
they're real and not "failed")