Selenium: Fix displayname test flakiness
WebDriverIO seems a bit fragile with it's async assertions, and it can
sometimes start executing the next test before the browser had time to
react to the previous test - in this case, before the edit console URL
modal was closed.

This caused flakiness with the edit display name test that comes
directly after the edit console URL test.

Waiting for the dropdown to be clickable (for up to 500ms, default
timeout) seems to fix this flakiness.