ClusterList: Don't crash when CS is down
When CS is down, OCM should still be usable, at least partially.

This commit makes it so when CS is down or returns any kind of error,
the cluster list will continue to function. OCP clusters will work
normally, and OSD clusters will show up, with unknown status, and with
all actions disabled (as if the user has no permission to edit/delete them).

Cluster Details is unaffected by this change.

https://issues.redhat.com/browse/SDA-3716