[SDA-3913] reset get schedules response when switching clusters
Resetting the response state when switching clusters, to avoid showing
incorrect data when switching between details of different clusters.

This is one of the downsides of doing *everything* in redux, a concept I
think we should reconsider - if this request was managed internally in
the component's state, we wouldn't have to reset it when unmounting.
same for every reset in the app.

I also used the same commit to fix an issue where some tab contents like
machinePools would appear below the Overview tab content when leaving
cluster details open while the cluster is installing.

https://issues.redhat.com/browse/SDA-3913