Remove xyaml/cucumber tests
These tests are slow, difficult to debug, and prone to strange errors.
Since it expects both an empty state and a cluster list with actual
clusters - it doesn't actually test anything, because the selenium user
doesn't actually have unarchived clusters. The tests we already have for
cluster creation interact with the cluster creation button, and therefor
should be sufficient for testing empty state functionality.

Keeping these tests seems to have zero benefits, and we'll need to
rewrite tests for the cluster list itself (with clusters!) anyway.

So, here's an MR removing these obsolete tests.