Filter cluster by cluster_id
In addition to the recent external_cluster_id and display_name.

Motivation: existing use-case to search for AI Cluster ID, not only by
name or OCP (CS) Cluster ID.

Since all IDs are generated UUIDs, there should not be real conflict
introduced by that change.

When an Assisted Installer cluster creates a subscription, the
cluster_id is set to the AI Cluster ID.