Change /subscriptions to /quota
For summit routing changes. This includes a redirect from /subscriptions
for old menu compatibility so we're not relying on a specific push order
on the Insights side. Once Insights completes their routing changes this
redirect will be ignored as `/subscriptions` won't go to OCM and we'll
be able to remove it.