SDA-4034: Normalize metrics to avoid crashes
SDA-4034 shows a crash that happened when the AMS backend didn't return
metrics at all. In the past we've seen a lot of issues with missing
fields, so this MR introduces a metrics normalization function that adds
fields when they're missing, in hope that crashes like this one can be
avoided.

It also makes sure we take metrics from the subscription even for OSD
clusters in the cluster list, and not just for single clusters - for
consistency.