SDA-4056: Fix insights menu highlighting
The Insights Platform expects us to call appNavClick to select which
menu item should be highlighted. This needs to happen both in full-page
reload and for in-app-navigation.

This MR unifies the parameters for appNavClick into one location, uses
the correct event property for finding the menu item's target path (for
in-app navigation) and takes basename differences between `/openshift`
and `/beta/openshift` into account.

https://issues.redhat.com/browse/SDA-4056