AddOnsCard: avoid crash when cluster has no console URL
In some rare cases we can reach a point where the addon is reported as
ready, but the `cluster` as passed to the `AddOnsCard` has no console
URL.

This commit fixes the issue by being defensive and making sure
`cluster.console.url` is available before trying to use it. If the
`cluster` object has no `console` field, accessing
`cluster.console.url` would cause a crash due to attempting access to a
property of an undefined variable.