Refactor BillingModelSection
This refactor is one of the many steps on the way to implementing the
OSD wizard.

Specifically, `BillingModelRadioButtons` was always supposed to be a
form control implementing the redux-form Field pattern. When
marketplace & trial were added, they were put in another field inside
`BillingModelRadioButtons`, making it a `Field` that contains another
`Field`, and making it more complicated to swap out the flat radio
buttons for regular radio buttons.

I opted for doing this refactor in a separate MR to make it easier to
review and keep the amount of changes minimal.