Don't rely on flag for OSD Trial cloud provider
This page is linked directly from openshift.com/try and there is a bug
where the feature flags may load after the quota loads making it appear
as though the user doesn't have access to this feature. The simplest fix
here is to remove the feature flag dependency in this page. This is not
a risk since we can still turn off this feature in AMS cluster auth if
necessary