Block rendering until config and insights are ready
Before the webpack 5 migration, we didn't call the render function until
config was ready. The migration moved the config initalization to be
out-of-band, creating race conditions when the app was rendered before
config was ready.

This commit is an attempt to block rendering until everything is ready,
by making `AppEntry` a class component with a ready state. The ready
state only sets to `true` when config is loaded, and until then
`AppEntry` will render nothing.