MGMT-6435: Hide tabs for unistalled AI clusters
The changes are relevant for AI clusters only.

If the cluster is not-yet installed, none of the Monitoring or Advisory
tabs are rendered.

If the cluster is installed, both tabs are visible no matter if telemetry
has reported its data. That way the user can optionally see their empty
states with farther guiding.