[SDA-4185] Fix ClusterDetailsRedirector crash by removing uuid-validate
`uuid-validate` uses a Node-specific API `Buffer`, which doesn't exist
in browsers. In Webpack 4.x, this was automatically polyfilled by
webpack itself. In Webpack 5.x, these polyfills have to be manual.

`uuid-validate` was only used by `subscriptionActions.js`, so instead of
figuring out a polyfill for the Buffer API it was much simpler to just
remove the library entirely - we already have another uuid library in
our dependencies that provides the same functionality.

This way we get two birds with one stone: cleaning up a dependency *and*
fixing an urgent bug.

https://issues.redhat.com/browse/SDA-4185