BillingModelSection: Wizard preparations
This commit prepares BillingModelSection for the OSD Wizard.

When the `isWizard` prop is provided, BillingModelSection will render
the introduction text from the wizard mockups, and render the infra
selection as normal radio buttons instead of flat radio buttons.

In addition to rendering changes, quota queries for this component were
moved to its own `index.js`. This has two benefits: It's more clear where the values come
from and how they're used, and it allows the Wizard code to be cleaner,
passing less props down the component tree.