Set boolean addon params to required based on validations
Required OCM addon parameters only indicate that a value must be set, and
not, in the case of a checkbox, that it must be checked and therefore true.
This change uses the validations on the param to determine if the desired
behavour is infact to make it required in the UI form.