SDA-4365: Add AWS ARM pre-release page
This MR adds an AWS ARM pre-release page. Breaking our usual convoluted
tech-debt pattern of splitting each page into two components, the page
is a single component. In the future we should refactor all OCP pages to be
a single component and token handling should be done in the token button
itself, not the page.

The links used are for installation from x86 machines to ARM instances.
I'm not sure this is the intention behind the request in the JIRA ticket
but it seems like a good starting point.

https://issues.redhat.com/browse/SDA-4365