Use separate installer tools for each cluster architecture
Installers are embedded with the payload from which they were extracted.
Because payloads are architecture specific, that means the installers
aren't by default interchangeable.  For instance, the Mac installer from
an x86_64 payload and the Mac installer from an ARM payload should not be
treated as the same.  However, clients are not payload dependent, and
therefore there is no issue using a client extracted from an x86_64 payload
with e.g. an IBM Z cluster of the same channel (version).

Also, the addition of an x86_64 installer to the non-x86_64 payloads,
along with the potential of adding Mac ARM binaries in some future version,
means that being able to select only the OS and not the architecture is
insufficient.  Therefore, this treats each payload architecture's installer
as a separate tool.