Fix quota cost any query with negative quota item
In many places the UI supplies "any" for various fields of the quota
cost query. This is problematic on an account with existing negative
quota. Negative quota is often granted to disable subscription based
legacy quota (for openshift v3). When making a quota query we should
not add to total available negative quota. This could cancel out valid
available quota when searching with "any" selectors. The unit test
covers this case