TrnasferClusterOwnership: Don't crash when subscription is undefined
When the user clicks Initiate Transfer, the modal data in the reducer is
cleared. Then immediately the component tries to render again.
`subscription` will be undefined on that render cycle, and if we don't
account for this it can cause a crash before ConnectedModal has a chance
to re-render and unmount the TransferClusterOwnershipDialog.

This commit fixes the issue by using the `?.` operator.