Make the `noproxy` mode default for `yarn start`
The `noproxy` mode, replacing the insights-proxy with a few lines of
webpack configuration, proved to be more stable and reliable than the
alternative.

It has been optional for a while now, it's time to make it the default
for `yarn start`. insights-proxy is still available to be started
manually for those who need it.