Selenium Downloads.js: await all expect()s
This commit only converts Downloads.js tests, rest deferred to future.

According to https://webdriver.io/docs/assertion/,
in async mode expect() returns a promise and has to be awaited as usual.

Without it, I believe in a sequence like this:

    expect(A).to...;
    await action();
    expect(B).to...;

there is a race between the action and expectation B, which might explain
some of the flakiness we see.

Also, I've confirmed that a test may pass despite expect(B) failing
(especially for last expectation in a test case)!