SDA-4648: Pass isTrialDefault to all wizard screens with initial values
This should fix SDA-4648 and allow creating a trial cluster via the
"trial default" approach (ie. clicking "Create Trial Cluster" and not
touching the billing model selection).

While I do wonder why do we need the duplicate functionality of having a
specific button for this (instead of just having the user select the
billing model in the first user screen) - as long as this functionality
exists it has to work as intended :)