Pass trailing args from `yarn selenium-with-* ...` to wdio
Drops WDIO_ARGS env var kludge.
Upgrading to concurrently 7.2.0 allows to pass args cleanly thanks to
https://github.com/open-cli-tools/concurrently/pull/307

The trailing `--` args are to prevent concurrently itself
from responding to flags there.  For example, without it
`yarn selenium-with-vnc --help` prints help of concurrently;
with the `--` it passes the flag on and it prints help of `wdio`
which is more consistent.