selenium: yarn upgrade 'chromedriver@*' to current 102
The `"chromedriver": "latest",` in package.json was ineffective because
yarn.lock does hardcode it to specific version, and Chromedriver 100
no longer works with current Chrome 102 (I think it already broke with 101).

(Changing "latest" to "*" because `yarn upgrade chromedriver` replaces
"latest" with specific major version e.g. "^102.0.0" after which future
`yarn upgrade chromedriver` stay on that 102; whereas with "*" it keeps
the "*" so future `yarn upgrade chromedriver` will keep upgrading it.)

This is a stop-gap "fix", lets `yarn selenium-with-chromedriver` work
with current Chrome 102 but still hardcoded, long-term we need a solution
that picks correct version for chrome version that developer has...

future refs:
https://chromedriver.chromium.org/downloads/version-selection
https://github.com/angular/webdriver-manager/issues/441
https://github.com/yeongbin-jo/python-chromedriver-autoinstaller