Allow ?env=... config overrides in production
When we implemented ?env overrides for debugging, we put in an extra
check to suppress this in production "just in case", but we never had
a real security reason for that.

It was already allowed on prod-beta, only blocked on prod-stable, which
doesn't make much sense.

The only concern is, since the override is "sticky" until manually cleared,
if an external customer somehow clicks a link with ?env= override,
could they get confused and not realize they must "Go back to normal"?

But external users can't use the other backends anyway.
https://api.stage.openshift.com IS world-reachable (no VPN needed, unlike
staging UI), but the backend blocks non-`@redhat.com` accounts with 401 error.
=> So what they'll see is "Environment override active" banner +
"This page is temporarily unavailable" error.

IMHO that's fine (it'd only be confusing if they could get functioning
staging UI and proceed to use it without understanding the banner).