downloads: only show available architectures for selected OS
Previously, in the case where a download was available on multiple
architectures for one OS (usually Linux) and a subset (but more than
one) of those architectures on another OS (usually Mac), then the
architectures supported by the former but not by the latter were still
listed in the dropdown, but shaded and not choosable, and with a tooltip.
Instead, just don't display the architecture at all.