./deploy_info.mjs: Work with older `curl` versions
`--fail-with-body` was added in curl 7.76.0.  It allows more details if
script fails, but not all team has it, so will do with `--fail`.
Either way we see the URL and error code but we're losing the response body:

-    cmd: 'curl --silent --show-error --fail-with-body https://raw.githubusercontent.com/RedHatInsights/uhc-portal-frontend-deploy/qa-stable/app.info.jsonx',
-    stdout: '404: Not Found',
+    cmd: 'curl --silent --show-error --fail https://raw.githubusercontent.com/RedHatInsights/uhc-portal-frontend-deploy/qa-stable/app.info.jsonx',
+    stdout: '',
     stderr: 'curl: (22) The requested URL returned error: 404\n'

So no big loss, can always test the URL manually if more details is needed.

(The main reason I've seen it fail is not being on VPN. :-)