[HAC-359] KMSKeyRingSelect, KMSKeySelect: highlight error when 0 options
Normally errors are not shown until field is `touched`.
But these fields are required, and when we got GCP data and know we
have 0 key rings / keys to offer, there is no point waiting for `touched`.
The present field is actually disabled, so nothing user _could_ touch!

So force-show the "Field is required" error immediately.

Also, tweaked the placeholder texts to hint which previous fields
user has to manipulate.