DynamicSelect: fix bug where "Loading..." label could get stuck
When switching Key ring location such that Key ring is invalidated,
and therefore Key name choice is irrelevant (hasDependencies=false)
and should render blank, the previous label — "Loading..." tended to
stick indefinitely after data loading was done.  Now becomes blank.

(I'm not sure how exactly it happened, as the React rendering did change to

    <FormSelectOption isDisabled value="" label={undefined} />

but apparently the `undefined` was confusing patternfly-react and the
actual `<option>` element was not getting updated?)