deploy_info.mjs: Normalize all hashes by `git rev-parse --short`
We've seen now a case where `src_hash` that we'd pushed in app.info.json
was shorter than hashes git prints for local branches.
Which is possible if more commits were added, or user configured `core.abbrev`.

Note that it'll still be possible for particular hashes to show longer
when they have prefix collisions; but this prevents the frequent annoyance
when the *same hash* is shown differently at branch / pushed / live.