Upgrade css-loader to avoid postcss error message
On some node versions (?), Sam and I have seen this build error:

    Module build failed: Error [ERR_PACKAGE_PATH_NOT_EXPORTED]: Package subpath './package.json' is not defined by "exports" in /home/beni/ocm/uhc-portal/node_modules/postcss/package.json

(repeating hundreds of times, for each .scss file.)

Sam reported this issues with Node 17 & 16, worked with 14.19.0 (and v12.22 in Jenkins).
For me it did work on v16.14.0 but now broke on F37 with v18.12.1.

This was fixed in postcss 8.2.12 and later:
https://github.com/postcss/postcss/commit/8e6088f99ca8ae92f70d5207a8478cd22e6ec420
which is an indirect dependency, upgrading css-loader allowed us to pick
a sufficiently new version — at least in the build-affecting path.