mockdata/record-real-cluster.sh: Record cluster's account & org
Internal users with `UHCSupport` role can read clusters from other
organizations, as well those accounts & org info (including quota).

When `current_account` != `cluster.creator`, UI isn't offering the same
actions as it would to the customer, which makes sense.
But when recording a cluster for debugging, it is desirable to reproduce
customer's point of view.
(came handy for debugging HAC-3604 but not specific to it.)

It's harmless as ?env=mockdata mode won't support actually making changes
to the cluster.
As before, you probably don't want to commit the account & org results to
master (we already print a warning about that).