[tests] Fail on "Maximum update depth exceeded"
When React logs this:
  console.error
    Error: Maximum update depth exceeded. This can happen when a component repeatedly calls setState inside componentWillUpdate or componentDidUpdate. React limits the number of nested updates to prevent infinite loops.

I assume it does limit the recursion depth but I've seen my test keep
running "forever" anyway (maybe the recursion was branching?).
An infinite run is useless and the output is hard to read,
both locally and more so in CI, so better abort the test immediately.