OCMUI-2908, OCMUI-2910 Access Request tab errors (#92)
# Summary

This PR fixes two issues related to counts on the Access Request tab:
total pending on the tab itself, and the total requests in the
pagination row.

Jira ticket: [OCMUI-2908](https://issues.redhat.com/browse/OCMUI-2908),
[OCMUI-2910](https://issues.redhat.com/browse/OCMUI-2910)

# Details

There are two separate root issues. One is that when the access request
data has changed, there isn't a hook to update the tab information. This
has been added to `TabsRow.tsx`. The second is that the total number of
access requests returned from the API is not placed into Redux so it can
be used by pagination. While not ideal, the best place to put this
dispatch is as soon as the data is fetched. See
`useFetchAccessRequests.tsx`

In addition, I removed code that was not functional and added a
subscription to a couple of React Queries because it is used by the
props sent to the service making the call.

# How to Test

This isn't easy to test - but ideally you would create at least 11
pending access requests for an HCP cluster. This is documented
[here](https://github.com/RedHatInsights/uhc-portal/wiki/Setup-and-Test-Rosa-Lockbox-feature).
Then follow the exact steps in each bug.

The easier option is to mock the data. So I'm adding the directions on
how to do this with Tweak.

1. Import this tweak file [ tweak file
](https://github.com/user-attachments/files/18413248/tweak_ocmui-2908.json)
 into [Tweak](https://tweak-extension.com/)
2. Enable the "0 items" mock. Ensure that the "12 items" mock is not
enabled
3. Go to an HCP cluster and then to to the access request tab
4. Without refreshing the page, enable the "12 items mock" and disable
the "0 items mock"
5. Refresh the cluster data (not using the browser refresh)
6. Ensure that the tab now has a "12" next to the tab title
7. Ensure that the pagination shows a total of 12 items
8. Change the per page value to 10
9. Ensure that the next page button is enabled. (Note clicking on the
button will not cause the data change because of the mocking)

## Video of Twek mocks

### Before


https://github.com/user-attachments/assets/0c406bd1-27aa-49dc-b90c-ced28c5bb277

### After


https://github.com/user-attachments/assets/d232b83e-a742-46d9-a6d5-a7b994e1a805

     

# Reviews

See [OCM UI PR into Master/Main process
guide](https://docs.google.com/document/d/1utGXwyP63cViOyLR7T2R7eU5BoeNOKMf7MyqjY1VApo/)
for more information.

## Reviewer 1 < name >

- [ ] Reviewed code
- [ ] Verified unit tests were added/modified for changed logic
- [ ] Verified change locally in a browser (downloaded and ran code)
- [ ] Closed threads I started after the author made changes or added an
explanation

## Reviewer 2 < name >

- [ ] Reviewed code
- [ ] Verified unit tests were added/modified for changed logic
- [ ] Verified change locally in a browser (downloaded and ran code)
- [ ] Closed threads I started after the author made changes or added an
explanation

NOTE: The author of the PR will merge the PR.

## Author

Check the following before merging:

- [x] Unit tests have been created and/or modified
- [x] All above checkboxes for both reviewers have been checked
- [x] All CI tests have passed
- [x] Has at least 2 approvals, ready to merge
- [x] Verify the 'Squash and merge' option is selected before merging
the PR into master (Click on the little arrow on the right of the green
merge button, and choose the 'Squash and merge' option)


# Ticketing

<!-- state the ticket or tickets this PR pertains to, e.g. "closes
OCMUI-nnn, OCMUI-mmm".
note that "fixes", "closes" or "resolves" (case-insensitive) will
automatically
move the ticket(s) to "review" upon merge.
to avoid this, you can use e.g. "addresses OCMUI-nnn" -->