OCMUI-2802: fixed update btn being wrongly shown (#93)
# Summary

When cluster update is set to recurring it was showing Update button on
regionalized clusters.

Jira ticket: [OCMUI-2802](https://issues.redhat.com/browse/OCMUI-2802)

# Details

Could not replicate bug precisely, but during investigation found that
when creating regionalized cluster with recurring update strategy, it
would resolve to manual update. This was due to api call for upgrade
strategy was failing because of the region not being passed into api
call.

# How to Test

Simply creating a regionalized cluster with recurring update and
navigating to the cluster settings tab.
Confirm that update button is not shown when the recurring update is
selected.

PS: When manual update is selected and you change it to recurring, it
takes around 5 seconds for update button to disappear. I assume it is
due to the latency to Singapore.


# Screen Captures

| Before | After |
| --------------------------------------------------- |
--------------------------------------- |
| <!-- attach a "before" screenshot or video here -->
|![image](https://github.com/user-attachments/assets/6c09def6-0c32-42ce-bc7d-6032a2aa7395)|
             

# Reviews

See [OCM UI PR into Master/Main process
guide](https://docs.google.com/document/d/1utGXwyP63cViOyLR7T2R7eU5BoeNOKMf7MyqjY1VApo/)
for more information.

## Reviewer 1 @kdoberst 

- [ ] Reviewed code
- [ ] Verified unit tests were added/modified for changed logic
- [ ] Verified change locally in a browser (downloaded and ran code)
- [ ] Closed threads I started after the author made changes or added an
explanation

## Reviewer 2 @DcooperRH 

- [ ] Reviewed code
- [ ] Verified unit tests were added/modified for changed logic
- [ ] Verified change locally in a browser (downloaded and ran code)
- [ ] Closed threads I started after the author made changes or added an
explanation

NOTE: The author of the PR will merge the PR.

## Author

Check the following before merging:

- [ ] Unit tests have been created and/or modified
- [ ] All above checkboxes for both reviewers have been checked
- [ ] All CI tests have passed
- [ ] Has at least 2 approvals, ready to merge
- [ ] Verify the 'Squash and merge' option is selected before merging
the PR into master (Click on the little arrow on the right of the green
merge button, and choose the 'Squash and merge' option)


# Ticketing

<!-- state the ticket or tickets this PR pertains to, e.g. "closes
OCMUI-nnn, OCMUI-mmm".
note that "fixes", "closes" or "resolves" (case-insensitive) will
automatically
move the ticket(s) to "review" upon merge.
to avoid this, you can use e.g. "addresses OCMUI-nnn" -->